joined channel now compare hased password
